requirements-dev.txt
flake8
flask-debugtoolbar
flask-shell-ipython
ipdb
ipython
isort
pytest
pytest-flask
pytest-cov
flask_sqlalchemy
flask_bcrypt
flask_wtf
flask_migrate
mysql-connector-python
pymysql
sqlite-web
RPi.GPIO
pi-rc522
flask-migrate
sqlite-web